Move Share Selected Documents handlers to separate package
